feat(contests.py): add error checking for if a user adds no attachment

+++ contests/contests.py
@@ -67,22 +67,25 @@ class ContestsCog(commands.Cog):
         async with ctx.channel.typing():
             channel_id = await self.config.guild(ctx.guild).posting_channel()
             channel = ctx.guild.get_channel(channel_id)
-            tempfile = await ctx.message.attachments[0].read()
-            extension = mimetypes.guess_extension(ctx.message.attachments[0].content_type)
-            author = ctx.message.author.name
-            author_id = ctx.message.author.id
-            await ctx.message.delete()
-            filehash = hashlib.md5(tempfile)
-            filename = filehash.hexdigest()
-            complete_name = f"{filename}{extension}"
-            discordfile = discord.File(filename=complete_name, fp=(io.BytesIO(tempfile)))
-            await channel.send(content=filename, file=discordfile)
-            contests_database_temp = await self.config.guild(ctx.guild).contests_database()
-            if type(contests_database_temp) is not dict:
-                contests_database_temp = {}
-            contests_database_temp[filename] = {
-                "author": author,
-                "author_id": author_id,
-            }
-            await self.config.guild(ctx.guild).contests_database.set(contests_database_temp)
+            if ctx.message.attachments is not None:
+                tempfile = await ctx.message.attachments[0].read()
+                extension = mimetypes.guess_extension(ctx.message.attachments[0].content_type)
+                author = ctx.message.author.name
+                author_id = ctx.message.author.id
+                await ctx.message.delete()
+                filehash = hashlib.md5(tempfile)
+                filename = filehash.hexdigest()
+                complete_name = f"{filename}{extension}"
+                discordfile = discord.File(filename=complete_name, fp=(io.BytesIO(tempfile)))
+                await channel.send(content=filename, file=discordfile)
+                contests_database_temp = await self.config.guild(ctx.guild).contests_database()
+                if type(contests_database_temp) is not dict:
+                    contests_database_temp = {}
+                contests_database_temp[filename] = {
+                    "author": author,
+                    "author_id": author_id,
+                }
+                await self.config.guild(ctx.guild).contests_database.set(contests_database_temp)
+            else:
+                await channel.send(content="Submission failed. Please attach an image to your message.")
